move/#50: 디렉토리 네이밍 변경
