+++ app/src/main/java/com/paw/key/presentation/ui/course/walkreview/component/WalkReviewInfoHolder.kt
@@ -1,4 +1,4 @@
-package com.paw.key.presentation.ui.course.walkrecord.component
+package com.paw.key.presentation.ui.course.walkreview.component
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Row
