+++ app/src/main/java/com/paw/key/presentation/ui/course/walkreview/component/WalkReviewTextField.kt
@@ -1,4 +1,4 @@
-package com.paw.key.presentation.ui.course.walkrecord.component
+package com.paw.key.presentation.ui.course.walkreview.component
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
